+++ pages/roadmaps/new.tsx
@@ -1,16 +1,11 @@
-/*
-TODO: 今apiリクエストをRoadmapコンポーネントから呼び出していて、
-pagesからしか呼び出さないっていうのを無視しちゃってる
-
-*/
-import RoadmapComp from '~/components/roadmaps/RoadmapComp'
+import RoadmapForm from '~/components/roadmaps/RoadmapForm'
 import { useAuth } from '~/contexts/AuthContext'
 import { createLibrary } from '~/utils/libraries'
 import type { Library, Roadmap } from '$prisma/client'
 import { RoadmapCreateBody, RoadmapUpdateBody } from '~/server/types'
 import { createRoadmap, updateRoadmap } from '~/utils/roadmaps'
 
-const createRoadmapsPageNew = () => {
+const NewRoadmapsPage = () => {
   const auth = useAuth()
   const onCreateLibrary = (title: Library['title'], link?: Library['link']) =>
     createLibrary(auth?.token || '', title, link)
@@ -19,7 +14,7 @@ const createRoadmapsPageNew = () => {
   const onUpdateRoadmap = (id: Roadmap['id'], data: RoadmapUpdateBody) =>
     updateRoadmap(auth?.token || '', id, data)
   return (
-    <RoadmapComp
+    <RoadmapForm
       onCreateLibrary={onCreateLibrary}
       onCreateRoadmap={onCreateRoadmap}
       onUpdateRoadmap={onUpdateRoadmap}
@@ -27,4 +22,4 @@ const createRoadmapsPageNew = () => {
   )
 }
 
-export default createRoadmapsPageNew
+export default NewRoadmapsPage
